+++ src/Component/JsonSchema/Guesser/Validator/Any/EnumValidator.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Jane\Component\JsonSchema\Guesser\Validator\Any;
+
+use Jane\Component\JsonSchema\Guesser\Guess\ClassGuess;
+use Jane\Component\JsonSchema\Guesser\Guess\Property;
+use Jane\Component\JsonSchema\Guesser\Validator\ValidatorGuess;
+use Jane\Component\JsonSchema\Guesser\Validator\ValidatorInterface;
+use Jane\Component\JsonSchema\JsonSchema\Model\JsonSchema;
+use Symfony\Component\Validator\Constraints\Choice;
+
+class EnumValidator implements ValidatorInterface
+{
+    public function supports($object): bool
+    {
+        return $object instanceof JsonSchema && ((\is_array($object->getType()) ? \in_array('string', $object->getType()) : 'string' === $object->getType()) || null === $object->getType()) && null !== $object->getEnum();
+    }
+
+    /**
+     * @param JsonSchema          $object
+     * @param ClassGuess|Property $guess
+     */
+    public function guess($object, string $name, $guess): void
+    {
+        $guess->addValidatorGuess(new ValidatorGuess(Choice::class, [
+            'choices' => $object->getEnum(),
+            'message' => '"{{ value }}" is not part of the set of possible choices for this field: "{{ choices }}".',
+        ]));
+    }
+}