+++ src/Component/JsonSchema/Guesser/Validator/Array_/MinItemsValidator.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Jane\Component\JsonSchema\Guesser\Validator\Array_;
+
+use Jane\Component\JsonSchema\Guesser\Guess\ClassGuess;
+use Jane\Component\JsonSchema\Guesser\Guess\Property;
+use Jane\Component\JsonSchema\Guesser\Validator\ValidatorGuess;
+use Jane\Component\JsonSchema\Guesser\Validator\ValidatorInterface;
+use Jane\Component\JsonSchema\JsonSchema\Model\JsonSchema;
+use Symfony\Component\Validator\Constraints\Count;
+
+class MinItemsValidator implements ValidatorInterface
+{
+    public function supports($object): bool
+    {
+        return $object instanceof JsonSchema && (\is_array($object->getType()) ? \in_array('array', $object->getType()) : 'array' === $object->getType()) && null !== $object->getMinItems();
+    }
+
+    /**
+     * @param JsonSchema          $object
+     * @param ClassGuess|Property $guess
+     */
+    public function guess($object, string $name, $guess): void
+    {
+        $guess->addValidatorGuess(new ValidatorGuess(Count::class, [
+            'min' => $object->getMinItems(),
+            'minMessage' => 'This array has not enough values. It should have {{ limit }} values or more.',
+        ]));
+    }
+}