+++ src/Component/JsonSchema/Guesser/Validator/String_/MinLengthValidator.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Jane\Component\JsonSchema\Guesser\Validator\String_;
+
+use Jane\Component\JsonSchema\Guesser\Guess\ClassGuess;
+use Jane\Component\JsonSchema\Guesser\Guess\Property;
+use Jane\Component\JsonSchema\Guesser\Validator\ValidatorGuess;
+use Jane\Component\JsonSchema\Guesser\Validator\ValidatorInterface;
+use Jane\Component\JsonSchema\JsonSchema\Model\JsonSchema;
+use Symfony\Component\Validator\Constraints\Length;
+
+class MinLengthValidator implements ValidatorInterface
+{
+    public function supports($object): bool
+    {
+        return $object instanceof JsonSchema && ((\is_array($object->getType()) ? \in_array('string', $object->getType()) : 'string' === $object->getType()) || null === $object->getType()) && null !== $object->getMinLength();
+    }
+
+    /**
+     * @param JsonSchema          $object
+     * @param ClassGuess|Property $guess
+     */
+    public function guess($object, string $name, $guess): void
+    {
+        $guess->addValidatorGuess(new ValidatorGuess(Length::class, [
+            'min' => $object->getMinLength(),
+            'minMessage' => 'This value is too short. It should have {{ limit }} characters or more.',
+        ]));
+    }
+}