+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [JsonSchema] [GH#405](https://github.com/janephp/janephp/pull/405) Validation
+
 ### Fixed
 - [JsonSchema] [GH#595](https://github.com/janephp/janephp/pull/595) Fix "oneOf" with nullable